+++ cloud/src/main/java/com/project/cloud/domain/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.project.cloud.auth.domain;
+package com.project.cloud.domain.user.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
