+++ cloud/src/main/java/com/project/cloud/domain/user/repository/UserRepository.java
@@ -1,11 +1,10 @@
-package com.project.cloud.auth.repository;
+package com.project.cloud.domain.user.repository;
 
-import com.project.cloud.auth.domain.User;
+import com.project.cloud.domain.user.entity.User;
 import java.util.Optional;
 import org.springframework.data.repository.Repository;
 
 public interface UserRepository extends Repository<User, Long> {
-
     User save(User email);
     Optional<User> findByEmail(String email);
 }
