+++ cloud/src/main/java/com/project/cloud/global/annotation/LoginUser.java
@@ -1,4 +1,4 @@
-package com.project.cloud.auth.annotaion;
+package com.project.cloud.global.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
