+++ cloud/src/main/java/com/project/cloud/global/auth/controller/OAuth2Controller.java
@@ -1,6 +1,6 @@
-package com.project.cloud.auth.controller;
+package com.project.cloud.global.auth.controller;
 
-import com.project.cloud.auth.annotaion.LoginUser;
+import com.project.cloud.global.annotation.LoginUser;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
