+++ cloud/src/main/java/com/project/cloud/global/auth/infra/OAuth2LoginSuccessHandler.java
@@ -1,6 +1,6 @@
-package com.project.cloud.auth.infrastructure;
+package com.project.cloud.global.auth.infra;
 
-import com.project.cloud.auth.util.JwtParser;
+import com.project.cloud.global.auth.util.JwtParser;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
