+++ cloud/src/main/java/com/project/cloud/global/config/SecurityConfig.java
@@ -1,8 +1,8 @@
-package com.project.cloud.auth.config;
+package com.project.cloud.global.config;
 
-import com.project.cloud.auth.infrastructure.JwtAuthenticationFilter;
-import com.project.cloud.auth.infrastructure.OAuth2LoginSuccessHandler;
-import com.project.cloud.auth.service.AuthFacadeService;
+import com.project.cloud.global.auth.infra.JwtAuthenticationFilter;
+import com.project.cloud.global.auth.infra.OAuth2LoginSuccessHandler;
+import com.project.cloud.global.auth.service.AuthFacadeService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
